Suggested revisions for clarity (this solution works)
