+++ Hand.java
@@ -47,6 +47,7 @@ public class Hand {
          for (int i = 1; i < hand.size(); i++) {
             Card c1 = (Card)hand.elementAt(i);
             if ( c1.getSuit() < c.getSuit() ||
+            		// getValue() is not a method on BlackJackCard
                     (c1.getSuit() == c.getSuit() && c1.getValue() < c.getValue()) ) {
                 pos = i;
                 c = c1;
